+++ src/audio_stream_player_csound_channel.h
@@ -28,7 +28,7 @@ public:
     virtual int _get_loop_count() const override;
     virtual double _get_playback_position() const override;
     virtual void _seek(double p_time) override;
-    virtual int _mix(AudioFrame *p_buffer, double p_rate_scale, int p_frames) override;
+    virtual int _mix(AudioFrame *p_buffer, float p_rate_scale, int p_frames) override;
     virtual float _get_length() const;
     AudioStreamPlaybackCsoundChannel();
     ~AudioStreamPlaybackCsoundChannel();
