+++ src/audio_stream_player_csound_named_channel.cpp
@@ -23,7 +23,7 @@ void AudioStreamPlaybackCsoundNamedChannel::_seek(double p_time) {
     }
 }
 
-int AudioStreamPlaybackCsoundNamedChannel::_mix(AudioFrame *p_buffer, double p_rate, int p_frames) {
+int AudioStreamPlaybackCsoundNamedChannel::_mix(AudioFrame *p_buffer, float p_rate, int p_frames) {
     ERR_FAIL_COND_V(!active, 0);
     if (!active) {
         return 0;
